+++ CHANGELOG.md
@@ -14,6 +14,10 @@ This file documents all changes to Argus-frontend. This file is primarily meant 
 - Logo and favicon
 - Add seconds to timestamps in elements of the event feed in detailed incident view
 
+- Order of events and acknowledgements in feed in detailed incident view. Order is now oldest-first.
+
+
 
 ### Added
 - End time to the incident detailed view (for closed incidents only)
+